requirements.txt
requests
sqlalchemy>=1.3.22
sqlalchemy_utils>=0.36.8
pydantic>=1.7.3
fastapi>=0.63.0
starlette>=0.13.6
uvicorn>=0.13.3
python-dotenv>=0.15.0
psycopg2-binary>=2.8.6
spacy>=2.3.5
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz